navigator: extract TelemetryService and RouteAnimator
Step 6 of the controllers/ refactor (see docs/refactor-controllers.md):
split AppController.startListening into two focused services and own
their lifecycles via start()/stop().

- services/telemetry.ts: TelemetryService — owns the tRPC subscription
  and the TelemetryTimeline buffer. Dispatches each event to its
  matching store action; calls routeRenderer.renderActiveRoute on
  routeUpdate.
- services/route-animator.ts: RouteAnimator — owns the 500ms render
  loop. Samples the timeline, writes truckPoint, drives the follow-
  camera and player-marker animation via MapAdapter, and asks
  RouteRenderer to redraw the active step's progress gradient.

AppControllerImpl shrinks: startListening becomes a thin proxy that
constructs and starts the two services on first call (lazy because
the services need an AppStore reference, which arrives via this
method). forceRePair stops both services. Drops eight more imports
that were only used by the inlined logic.

Note: ControlsController still has its own duplicate tRPC subscription
for speed/limit. Dropping that is a separate commit (changes both
sides of the wire and ControlsStore's write path).
